+++ bin-src/lib/getDependentStoryFiles.test.ts
@@ -19,6 +19,7 @@ const getContext: any = ({
   options: { storybookBaseDir: '.', ...options },
   turboSnap: {},
   storybook: { configDir, staticDir },
+  git: {},
 });
 
 afterEach(() => {
@@ -185,6 +186,38 @@ describe('getDependentStoryFiles', () => {
     });
   });
 
+  it('detects indirect changes to CSF files, angular css resource', async () => {
+    const changedFiles = ['src/foo.css'];
+    const modules = [
+      {
+        id: null,
+        name: './src/foo.css?ngResource ',
+        reasons: [{ moduleName: './src/foo.component.ts' }],
+      },
+      {
+        id: './src/foo.component.ts',
+        name: './src/foo.component.ts + 1 modules',
+        modules: [{ name: './src/foo.component.ts' }, { name: './src/foo.css?ngResource' }],
+        reasons: [{ moduleName: './src/foo.stories.ts' }],
+      },
+      {
+        id: './src/foo.stories.ts',
+        name: './src/foo.stories.ts',
+        reasons: [{ moduleName: CSF_GLOB }],
+      },
+      {
+        id: CSF_GLOB,
+        name: CSF_GLOB,
+        reasons: [{ moduleName: './.storybook/generated-stories-entry.js' }],
+      },
+    ];
+    const ctx = getContext();
+    const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
+    expect(res).toEqual({
+      './src/foo.stories.ts': ['src/foo.stories.ts'],
+    });
+  });
+
   it('detects indirect changes to CSF files in a single module chunk', async () => {
     const changedFiles = ['src/foo.js'];
     const modules = [
@@ -308,7 +341,7 @@ describe('getDependentStoryFiles', () => {
   });
 
   it('bails on changed global file', async () => {
-    const changedFiles = ['src/foo.stories.js', 'src/package.json'];
+    const changedFiles = ['src/foo.stories.js', 'src/yarn.lock'];
     const modules = [
       {
         id: './src/foo.stories.js',
@@ -325,10 +358,39 @@ describe('getDependentStoryFiles', () => {
     const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
     expect(res).toEqual(null);
     expect(ctx.turboSnap.bailReason).toEqual({
+      changedPackageFiles: ['src/yarn.lock'],
+    });
+    expect(ctx.log.warn).toHaveBeenCalledWith(
+      expect.stringContaining(chalk`Found a package file change in {bold src/yarn.lock}`)
+    );
+  });
+
+  it('bails on dependency changes to package manifest file', async () => {
+    const changedFiles = ['src/foo.stories.js', 'src/package.json'];
+    const modules = [
+      {
+        id: './src/foo.stories.js',
+        name: './src/foo.stories.js',
+        reasons: [{ moduleName: CSF_GLOB }],
+      },
+      {
+        id: CSF_GLOB,
+        name: CSF_GLOB,
+        reasons: [{ moduleName: './.storybook/generated-stories-entry.js' }],
+      },
+    ];
+    const rawContext = getContext();
+    const ctx = {
+      ...rawContext,
+      git: { ...rawContext.git, changedPackageManifests: ['src/package.json'] },
+    };
+    const res = await getDependentStoryFiles(ctx, { modules }, statsPath, changedFiles);
+    expect(res).toEqual(null);
+    expect(ctx.turboSnap.bailReason).toEqual({
       changedPackageFiles: ['src/package.json'],
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
-      expect.stringContaining(chalk`Found a package file change in {bold src/package.json}`)
+      expect.stringContaining(chalk`Found a dependency change in {bold src/package.json}`)
     );
   });
 
@@ -354,7 +416,7 @@ describe('getDependentStoryFiles', () => {
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(
-        chalk`Found a Storybook config file change in {bold path/to/storybook-config/file.js}`
+        chalk`Found a Storybook config change in {bold path/to/storybook-config/file.js}`
       )
     );
   });
@@ -381,7 +443,7 @@ describe('getDependentStoryFiles', () => {
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(
-        chalk`Found a Storybook config file change in {bold path/to/storybook-config/file.js}`
+        chalk`Found a Storybook config change in {bold path/to/storybook-config/file.js}`
       )
     );
   });
@@ -409,7 +471,7 @@ describe('getDependentStoryFiles', () => {
     });
     expect(ctx.log.warn).toHaveBeenCalledWith(
       expect.stringContaining(
-        chalk`Found a Storybook config file change in {bold path/to/storybook-config/file.js}`
+        chalk`Found a Storybook config change in {bold path/to/storybook-config/file.js}`
       )
     );
   });