+++ bin-src/tasks/gitInfo.ts
@@ -6,7 +6,7 @@ import { getParentCommits } from '../git/getParentCommits';
 import { getBaselineBuilds } from '../git/getBaselineBuilds';
 import { exitCodes, setExitCode } from '../lib/setExitCode';
 import { createTask, transitionTo } from '../lib/tasks';
-import { matchesFile } from '../lib/utils';
+import { matchesFile, isPackageManifestFile } from '../lib/utils';
 import {
   initial,
   pending,
@@ -23,7 +23,7 @@ import { Context, Task } from '../types';
 import { getChangedFilesWithReplacement } from '../git/getChangedFilesWithReplacement';
 import replacedBuild from '../ui/messages/info/replacedBuild';
 import forceRebuildHint from '../ui/messages/info/forceRebuildHint';
-import arePackageDependenciesEqual from '../lib/comparePackageJsons';
+import { arePackageDependenciesEqual } from '../lib/comparePackageManifests';
 
 const SkipBuildMutation = `
   mutation SkipBuildMutation($commit: String!, $branch: String, $slug: String) {
@@ -51,6 +51,23 @@ interface LastBuildQueryResult {
     };
   };
 }
+
+const getPackageManagerChanges = async (
+  commit: string,
+  changedPackageFiles: string[]
+): Promise<string[]> => {
+  const allChanges = await Promise.all(
+    changedPackageFiles.map(async (fileName) => {
+      const fileA = await execGitCommand(`git show ${commit}:${fileName}`);
+      const fileB = await execGitCommand(`git show HEAD:${fileName}`);
+
+      // put in empty entry for equal-dependency packages so we only have fileNames for the non-equal ones
+      return arePackageDependenciesEqual(JSON.parse(fileA), JSON.parse(fileB)) ? [] : [fileName];
+    })
+  );
+
+  return allChanges.flat();
+};
 
 export const setGitInfo = async (ctx: Context, task: Task) => {
   const { branchName, patchBaseRef, fromCI: ci, interactive } = ctx.options;
@@ -162,6 +179,22 @@ export const setGitInfo = async (ctx: Context, task: Task) => {
         }))
       );
       ctx.git.changedFiles = Array.from(new Set(results.flatMap((r) => r.changedFiles)));
+
+      const packageManifestChanges: { commit: string; changedFiles: string[] }[] = [];
+
+      results.forEach((resultItem) => {
+        const changedPackageFiles = resultItem.changedFiles.filter((changedFile) =>
+          isPackageManifestFile(changedFile)
+        );
+
+        if (changedPackageFiles.length) {
+          packageManifestChanges.push({
+            commit: resultItem.build.commit,
+            changedFiles: changedPackageFiles,
+          });
+        }
+      });
+
       ctx.git.replacementBuildIds = results
         .filter((r) => !!r.replacementBuild)
         .map(({ build, replacementBuild }) => {
@@ -177,16 +210,15 @@ export const setGitInfo = async (ctx: Context, task: Task) => {
         );
       }
 
-      const changedPackageFiles = ctx.git.changedFiles.filter((fileName) =>
-        [/^package\.json$/, /\/package\.json$/].some((re) => re.test(fileName))
+      const packageManifestDependencyChanges = await Promise.all(
+        packageManifestChanges.map(async ({ commit: buildCommit, changedFiles }) => {
+          return getPackageManagerChanges(buildCommit, changedFiles);
+        })
       );
 
-      if (changedPackageFiles.length > 0) {
-        ctx.git.packageControlFilesWithDependencyChanges = await getPackageManagerChanges(
-          ctx.build,
-          changedPackageFiles
-        );
-      }
+      const flattenedChanges = packageManifestDependencyChanges.flat();
+      // remove duplicate entries (if have multiple ancestors and both changed the same package.json, for example)
+      ctx.git.changedPackageManifests = Array.from(new Set(flattenedChanges));
     } catch (e) {
       ctx.turboSnap.bailReason = { invalidChangedFiles: true };
       ctx.git.changedFiles = null;
@@ -209,26 +241,6 @@ export const setGitInfo = async (ctx: Context, task: Task) => {
   }
 
   transitionTo(success, true)(ctx, task);
-};
-
-const getPackageManagerChanges = async (
-  // TODO: type of build
-  build,
-  changedPackageFiles: string[]
-): Promise<string[]> => {
-  const allChanges = await Promise.all(
-    changedPackageFiles.map(async (fileName) => {
-      const fileA = await execGitCommand(`git show ${build.commit}:${fileName}`);
-      const fileB = await execGitCommand(`git show HEAD:${fileName}`);
-
-      return {
-        fileName,
-        sameDependencies: arePackageDependenciesEqual(JSON.parse(fileA), JSON.parse(fileB)),
-      };
-    })
-  );
-
-  return allChanges.filter((change) => !change.sameDependencies).map((change) => change.fileName);
 };
 
 export default createTask({